requirements.txt
google-cloud-secret-manager
google-cloud-storage
requests
google-cloud-secret-manager
requests